+++ packages/api-client/src/layouts/Modal/create-api-client-modal.ts
@@ -31,18 +31,20 @@ export const createApiClientModal = async ({
 
   const { importSpecFile, importSpecFromUrl } = client.store
 
-  // Import the spec if needed
-  if (configuration.spec?.url)
-    await importSpecFromUrl(configuration.spec.url, 'default', {
-      proxyUrl: configuration.proxyUrl,
-      setCollectionSecurity: true,
-      ...configuration,
-    })
-  else if (configuration.spec?.content)
-    await importSpecFile(configuration.spec?.content, 'default', {
-      setCollectionSecurity: true,
-      ...configuration,
-    })
+  // Import the spec if we don't pass in a store
+  if (!store) {
+    if (configuration.spec?.url)
+      await importSpecFromUrl(configuration.spec.url, 'default', {
+        proxyUrl: configuration.proxyUrl,
+        setCollectionSecurity: true,
+        ...configuration,
+      })
+    else if (configuration.spec?.content)
+      await importSpecFile(configuration.spec?.content, 'default', {
+        setCollectionSecurity: true,
+        ...configuration,
+      })
+  }
 
   return client
 }
