+++ packages/api-client/src/store/import-spec.ts
@@ -70,6 +70,8 @@ export function importSpecFileFactory({
       workspaceEntities.collection.uid,
     ])
 
+    console.timeEnd('populate workspace')
+
     return workspaceEntities.collection
   }
