+++ packages/api-reference/src/components/Content/Lazy/Loading.vue
@@ -169,8 +169,8 @@ onMounted(() => {
           :collection="activeCollection"
           :layout="layout"
           :operation="operation"
-          :request="getRequest(requests, operation.path, operation.httpVerb)"
           :requestExamples="requestExamples"
+          :requests="requests"
           :securitySchemes="securitySchemes"
           :server="activeServer"
           :tag="tag" />
