+++ packages/api-reference/src/components/Content/Tag/TagList.vue
@@ -1,7 +1,6 @@
 <script setup lang="ts">
 import { Lazy } from '@/components/Content/Lazy'
 import { Operation } from '@/features/Operation'
-import { getRequest } from '@/helpers/get-request'
 import { useNavState, useSidebar } from '@/hooks'
 import { useActiveEntities, useWorkspace } from '@scalar/api-client/store'
 import { ScalarErrorBoundary } from '@scalar/components'
@@ -50,8 +49,8 @@ const isLazy = props.layout !== 'classic' && !hash.value.startsWith('model')
             :collection="activeCollection"
             :layout="layout"
             :operation="operation"
-            :request="getRequest(requests, operation.path, operation.httpVerb)"
             :requestExamples="requestExamples"
+            :requests="requests"
             :securitySchemes="securitySchemes"
             :server="activeServer"
             :tag="tag" />
